test(unit): split the last remaining unittest-based classes into modules"
